requirements.txt
amqp==5.1.1
apturl==0.5.2
asgiref==3.5.2
attrs==21.2.0
autobahn==22.4.2
Automat==20.2.0
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
Brlapi==0.8.3
celery==5.2.7
certifi==2020.6.20
cffi==1.15.0
channels==3.0.4
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
command-not-found==0.3
constantly==15.1.0
cryptography==37.0.2
cupshelpers==1.0
daphne==3.0.2
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20200713
distro==1.7.0
distro-info===1.1build1
Django==3.2.13
django-allauth==0.50.0
django-celery-beat==2.2.1
django-elasticsearch-dsl==7.2.2
django-import-export==2.8.0
django-model-utils==4.2.0
django-notifications-hq==1.7.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
elasticsearch==7.17.3
elasticsearch-dsl==7.4.0
et-xmlfile==1.1.0
fasteners==0.14.1
future==0.18.2
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
import-export==0.2.67.dev6
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
jsonfield==3.1.0
jsonschema==3.2.0
kazam==1.4.5
keyring==23.5.0
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupPy==1.14
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.10
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
prompt-toolkit==3.0.29
protobuf==3.12.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
PyGObject==3.42.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyotp==2.6.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.3.0+ubuntu2
python-crontab==2.6.0
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
SecretStorage==3.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
stripe==3.2.0
swapper==1.3.0
systemd-python==234
tablib==3.2.1
texttable==1.6.4
twilio==7.9.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vine==5.0.0
wadllib==1.3.6
wcwidth==0.2.5
websocket-client==1.2.3
xdg==5
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
zipp==1.0.0
zope.interface==5.4.0
